Changed TacticalPing so that it uses custom strings
The reason for this change is because of a hidden detail about the tactical ping parameter. When the option is assigned one or two, it will ONLY show up inside the 3D view of the game. You will not be able to view the tactical ping on the map unless you set TacticalPing to three. This change hopes to bring clarity to this hidden detail and provide comfort and insight for the user.